+++ spec/dummy/db/migrate/20150825231046_add_slug_to_aromas.rb
@@ -1,0 +1,5 @@
+class AddSlugToAromas < ActiveRecord::Migration
+  def change
+    add_column :aromas, :slug, :string
+  end
+end